// backend/src/mysql/models/Field_names.ts
export class Field_names{
    idNazivi_polja: number = 0;
    polje1: string = "";
    polje2: string = "";
    polje3: string = "";
    polje4: string = "";
    polje5: string = "";
    polje6: string = "";
    polje7: string = "";
    polje8: string = "";
    polje9: string = "";
    polje10: string = "";
    polje11: string = "";
    polje12: string = "";
}